+++ engines/order_management/app/services/reports.rb
@@ -1,0 +1,3 @@
+module Reports
+  class UnsupportedReportFormatException < StandardError; end
+end
