+++ engines/order_management/app/services/reports/authorizer.rb
@@ -1,0 +1,12 @@
+module Reports
+  class Authorizer
+    class ParameterNotAllowedError < StandardError; end
+
+    attr_accessor :parameters, :permissions
+
+    def initialize(parameters, permissions)
+      @parameters = parameters
+      @permissions = permissions
+    end
+  end
+end
