+++ engines/order_management/app/services/reports/permissions.rb
@@ -1,0 +1,9 @@
+module Reports
+  class Permissions
+    attr_accessor :user
+
+    def initialize(user)
+      @user = user
+    end
+  end
+end
