+++ engines/order_management/app/services/reports/renderers/base.rb
@@ -1,0 +1,11 @@
+module Reports
+  module Renderers
+    class Base
+      attr_reader :report_data
+
+      def initialize(report_data)
+        @report_data = report_data
+      end
+    end
+  end
+end
