+++ engines/order_management/app/services/reports/report_data/base.rb
@@ -1,0 +1,11 @@
+module Reports
+  module ReportData
+    class Base
+      def initialize(attributes = {})
+        attributes.each do |key, value|
+          public_send("#{key}=", value)
+        end
+      end
+    end
+  end
+end
