+++ engines/order_management/lib/order_management/reports/enterprise_fee_summary/report_data/enterprise_fee_type_total.rb
@@ -1,10 +1,8 @@
-require "open_food_network/reports/report_data/base"
-
 module OrderManagement
   module Reports
     module EnterpriseFeeSummary
       module ReportData
-        class EnterpriseFeeTypeTotal < OpenFoodNetwork::Reports::ReportData::Base
+        class EnterpriseFeeTypeTotal < ::Reports::ReportData::Base
           attr_accessor :fee_type, :enterprise_name, :fee_name, :customer_name, :fee_placement,
                         :fee_calculated_on_transfer_through_name, :tax_category_name, :total_amount
