+++ engines/order_management/lib/order_management/reports/enterprise_fee_summary/report_data/enterprise_fee_type_totals.rb
@@ -1,10 +1,8 @@
-require "open_food_network/reports/report_data/base"
-
 module OrderManagement
   module Reports
     module EnterpriseFeeSummary
       module ReportData
-        class EnterpriseFeeTypeTotals < OpenFoodNetwork::Reports::ReportData::Base
+        class EnterpriseFeeTypeTotals < ::Reports::ReportData::Base
           attr_accessor :list
 
           def initialize(*args)
