+++ engines/order_management/spec/lib/order_management/reports/enterprise_fee_summary/parameters_spec.rb
@@ -77,7 +77,7 @@ describe OrderManagement::Reports::EnterpriseFeeSummary::Parameters do
 
     it "raises error when the parameters are not allowed" do
       parameters = described_class.new(order_cycle_ids: [(order_cycle.id + 1).to_s])
-      expect { parameters.authorize!(permissions) }.to raise_error(OpenFoodNetwork::Reports::Authorizer::ParameterNotAllowedError)
+      expect { parameters.authorize!(permissions) }.to raise_error(Reports::Authorizer::ParameterNotAllowedError)
     end
   end
