--- lib/open_food_network/reports.rb
@@ -1,5 +1,0 @@
-module OpenFoodNetwork
-  module Reports
-    class UnsupportedReportFormatException < StandardError; end
-  end
-end
