--- lib/open_food_network/reports/authorizer.rb
@@ -1,14 +1,0 @@
-module OpenFoodNetwork
-  module Reports
-    class Authorizer
-      class ParameterNotAllowedError < StandardError; end
-
-      attr_accessor :parameters, :permissions
-
-      def initialize(parameters, permissions)
-        @parameters = parameters
-        @permissions = permissions
-      end
-    end
-  end
-end
