--- lib/open_food_network/reports/permissions.rb
@@ -1,11 +1,0 @@
-module OpenFoodNetwork
-  module Reports
-    class Permissions
-      attr_accessor :user
-
-      def initialize(user)
-        @user = user
-      end
-    end
-  end
-end
