--- lib/open_food_network/reports/renderers/base.rb
@@ -1,13 +1,0 @@
-module OpenFoodNetwork
-  module Reports
-    module Renderers
-      class Base
-        attr_reader :report_data
-
-        def initialize(report_data)
-          @report_data = report_data
-        end
-      end
-    end
-  end
-end
