--- lib/open_food_network/reports/report_data/base.rb
@@ -1,13 +1,0 @@
-module OpenFoodNetwork
-  module Reports
-    module ReportData
-      class Base
-        def initialize(attributes = {})
-          attributes.each do |key, value|
-            public_send("#{key}=", value)
-          end
-        end
-      end
-    end
-  end
-end
